Silence pep8 E402
pep8 1.6.0 adds a new error E402, which complains when a module level
import is made somewhere other than at the top of the file. This new
addition is causing bin/lint.bash to fail, and thus our automated CI
tests to fail.

This fix is a stopgap, allowing our automated CI tests to pass by
silencing E402. If we later decide we agree with E402, and make the
necessary corrections, we should re-enable E402.