doap: remove category
The "desktop" category was rejected by the pre-hooks in the GNOME repo.  Based
on bugzilla, MouseTrap is currently in the "other" category.  But there doesn't
appear to be an "other" category for doaps to use. The alternative appears to be
to leave it out. So I'm deleting the category for now.