Shutdown GTK on SIGINT and SIGTERM
Currently MouseTrap hangs when a SIGINT (e.g., ^C from the terminal)
or a SIGTERM (e.g., kill PID) is received. The reason is that the
main event loop in GTK is still running.

This commit adds a signal handler to respond to SIGINT and SIGTERM,
and adds code to shutdown GTK.

Closes bug 750077