Move Loop to vision.py
Ultimately the CameraComponent is what will have the "loop", the
"clock", the "pulse". It will capture an image from the webcam
at regular intervals, and then fire an event signaling the other
components that an image has been captured.

So, this commit moves Loop to vision.py where it will be integrated
into CameraComponent.