Use correct python version for `make check` and `make coverage`
Before, `make check` did one of two things:

1) It ran the tests indirectly using `coverage`.
2) It ran the tests directly.

It did the first if `coverage` was available, otherwise it did
the second.

The problem was that `coverage` always used python2, whereas the
second used whatever python version was specified when autogen.sh
was run. This gave inconsistant test results.

This commit simplifies `make check` so that it always runs the
tests directly using the version of python specified with
autogen.sh.

The call to coverage is now isolated to `make coverage`. Also,
coverage is called using a specific python version: the version
specified with autogen.sh.

This should give more consistent results.

Closes bug 751066